feat: 修改条码删除

+++ HymsonMES/src/Libraries/Hymson.MES.Services/Services/Plan/PlanSfcPrint/PlanSfcPrintService.cs
@@ -16,6 +16,7 @@ using Hymson.MES.CoreServices.Services.Common.ManuCommon;
 using Hymson.MES.CoreServices.Services.Manufacture.ManuCreateBarcode;
 using Hymson.MES.Data.Repositories.Common.Command;
 using Hymson.MES.Data.Repositories.Manufacture;
+using Hymson.MES.Data.Repositories.Manufacture.ManuSfc.Command;
 using Hymson.MES.Data.Repositories.Plan;
 using Hymson.MES.Data.Repositories.Process;
 using Hymson.MES.HttpClients;
@@ -25,6 +26,7 @@ using Hymson.MES.Services.Dtos.Plan;
 using Hymson.Snowflake;
 using Hymson.Utils;
 using Hymson.Utils.Tools;
+using IdGen;
 
 namespace Hymson.MES.Services.Services.Plan
 {
@@ -256,7 +258,7 @@ namespace Hymson.MES.Services.Services.Plan
         /// </summary>
         /// <param name="idsArr"></param>
         /// <returns></returns>
-        public async Task<int> DeletesAsync(long[] idsArr)
+        public async Task<int> DeletesAsync(IEnumerable<long> idsArr)
         {
             var sfcEntities = await _manuSfcRepository.GetByIdsAsync(idsArr);
             if (sfcEntities.Any(it => it.IsUsed == YesOrNoEnum.Yes)) throw new CustomerValidationException(nameof(ErrorCode.MES16116));
@@ -265,22 +267,34 @@ namespace Hymson.MES.Services.Services.Plan
             // 对锁定状态进行验证
             await _manuCommonService.VerifySfcsLockAsync(new ManuProcedureBo
             {
-                SiteId = 0,
+                SiteId = _currentSite.SiteId ?? 0,
                 SFCs = sfcEntities.Select(s => s.SFC)
             });
 
             // 条码集合
             var sfcInfoEntities = await _manuSfcInfoRepository.GetBySFCIdsAsync(sfcEntities.Select(s => s.Id));
 
+            var manuSfcUpdateStatusByIdCommands = new List<ManuSfcUpdateStatusByIdCommand>();
+
+            foreach (var item in sfcEntities)
+            {
+                manuSfcUpdateStatusByIdCommands.Add(new ManuSfcUpdateStatusByIdCommand
+                {
+                    Id = item.Id,
+                    Status = SfcStatusEnum.Delete,
+                    CurrentStatus = item.Status,
+                    UpdatedBy = _currentUser.UserName
+                });
+            }
+
             var rows = 0;
             using (var trans = TransactionHelper.GetTransactionScope())
             {
-                rows += await _manuSfcRepository.DeletesAsync(new DeleteCommand
+                rows += await _manuSfcRepository.ManuSfcUpdateStatuByIdRangeAsync(manuSfcUpdateStatusByIdCommands);
+                if (rows != manuSfcUpdateStatusByIdCommands.Count)
                 {
-                    Ids = idsArr,
-                    UserId = _currentUser.UserName,
-                    DeleteOn = HymsonClock.Now()
-                });
+                    throw new CustomerValidationException(nameof(ErrorCode.MES16138));
+                }
 
                 rows += await _manuSfcProduceRepository.DeletePhysicalRangeAsync(new Data.Repositories.Manufacture.ManuSfcProduce.Command.DeletePhysicalBySfcsCommand()
                 {
